refactor: move utility functions into separate module
